==> scripts/2_Particion_datos_externa.ipynb <==
import pandas as pd
import numpy as np

# Leer la información del documento Excel original
archivo = '../database.csv'

df = pd.read_csv(archivo, sep=';')

print(df)

#Particion externa tipo hold out 80-20%
from sklearn.model_selection import train_test_split
train, test = train_test_split(df, test_size=0.2, shuffle=True, random_state=42) # hold-out

# Aleatorizar los dataframes
train = train.sample(frac=1)
test = test.sample(frac=1)


# Guardado de los modelos
import os
if not os.path.exists('../partitions'):
    os.mkdir('../partitions')
train.to_csv('../partitions/train.csv', sep=';')
test.to_csv('../partitions/test.csv', sep=';')
