==> scripts/5_Modelado.ipynb <==
# Cargar los datos de entrenamiento
import numpy as np
train = np.load('../final_features/train.npy')

X_train = train[:,:-1]
y_train = train[:, -1]

print(np.shape(X_train))
print(np.shape(y_train))
import sklearn
print(sklearn.__version__)

# Definir los modelos de clasificación
from sklearn.linear_model import LogisticRegression
from sklearn.neural_network import MLPClassifier
from sklearn.model_selection import cross_val_score, Kfold

algoritmos = {'LOGR': LogisticRegression(penalty='l2', solver='saga', max_iter=1000, random_state=42),
             'MLP': MLPClassifier(hidden_layer_sizes=[8,4], activation='relu', solver='sgd', batch_size='auto',
                                  learning_rate='adaptive', learning_rate_init=0.01, max_iter=1000, random_state=42)}

# Cross-validation interno en k=5 bolsas, no he conseguido arreglarlo debe de ser relacionado con la versión


results={}
for nombre, alg in algoritmos.items():
    results[nombre] = cross_val_score(alg, X_train, y_train, cv=Kfold(n_splits=5, shuffle=True, random_state=42))
    print(nombre + ':  Accuracy:  %0.4f +/- %0.4f'% (results[nombre].mean(), results[nombre].std()))

# Definimos el modelo definitivo.
algoritmos = {'LOGR': LogisticRegression(penalty='l2', solver='saga', max_iter=10000, random_state=42),
             'MLP': MLPClassifier(hidden_layer_sizes=[200,100], activation='relu', solver='adam', batch_size='auto',
                                  learning_rate='adaptive', learning_rate_init=0.001, max_iter=1000, random_state=42)}

LOGR_definitivo = algoritmos['LOGR'].fit(X_train, y_train)
MLP_definitivo = algoritmos['MLP'].fit(X_train, y_train)

# Atributos se obtienen durante el entrenamiento
print('Mínimo error cometido: ', MLP_definitivo.best_loss_)
print('Número de iteraciones llevadas a cabo: ', MLP_definitivo.n_iter_)

Como podemos ver la mínimo error es muy alto, aunque mejoremos el clasificador a partir de los hiperparametros segurmanete, es tan mal definidos la selección de caracteres, aunque seguiremos para adelante con estos modelos

# Guardar modelos
import os
if not os.path.exists('../models'):
    os.mkdir('../models')

import pickle
with open('../models/LOGR.pickle', 'wb') as fw:
    pickle.dump(LOGR_definitivo, fw)
with open('../models/MLP.pickle', 'wb') as fw:
    pickle.dump(MLP_definitivo, fw)
