==> scripts/6_Evaluación.ipynb <==
# Cargar los datos del test
import numpy as np
test = np.load('../final_features/test.npy')
#quitamos la última columna que es la del identificador
X_test = test[:,:-1]
y_test = test[:, -1]

print(np.shape(X_test))
print(np.shape(y_test))

# Cargamos los modelos entrenados
import pickle
with open('../models/LOGR.pickle', 'rb') as fr:
    LOGR = pickle.load(fr)
with open('../models/MLP.pickle', 'rb') as fr:
    MLP = pickle.load(fr)

# Extraer las predicciones
y_pred_LOGR = LOGR.predict(X_test)
y_pred_MLP = MLP.predict(X_test)

y_pred = [y_pred_LOGR, y_pred_MLP]

# Evaluar diversas métricas de clasificación
from sklearn import metrics

from tabulate import tabulate
headers = ['', 'LOGR', 'MPL']
P, S, FS, ACC, AUC = [['Precision'], ['Sensibilidad'], ['F1-Score'], ['Accuracy'], ['AUC']]

for i in range(0,2):
    P.append(np.round(metrics.precision_score(y_test, y_pred[i]),4))
    S.append(np.round(metrics.recall_score(y_test, y_pred[i]),4))
    FS.append(np.round(metrics.f1_score(y_test, y_pred[i]),4))
    ACC.append(np.round(metrics.accuracy_score(y_test, y_pred[i]),4))
    AUC.append(np.round(metrics.roc_auc_score(y_test, y_pred[i]),4))

my_data = [tuple(P), tuple(S), tuple(FS), tuple(ACC), tuple(AUC)]
print(tabulate(my_data, headers=headers))

# Confusion matrix
print(metrics.confusion_matrix(y_test, y_pred[1]))

Como conclusión a los dos modelos, el clasificador MPL ha obtenido mejores resultados que el LOGR, como era de esperar, los resultados no son muy buenos, ya que seguramente la obtención de la extracción de características no se ha realizado correctamente, para un futuro se podrían mejorar estos resultados.
